Created separate test file for create command tests
